+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix bug concerning table and separator alignment of multi-line hash with multiple keys on the same line.
 * Fix a bug where `ClassLength` counted lines of inner classes/modules
 * Fix a false positive for namespace class in `Documentation`
+* Fix "Parser::Source::Range spans more than one line" bug in clang formatter
 
 ## 0.14.0 (07/10/2013)